I refactor location of getZoraProfileScore.
